Work on finding roles which can be signed by the specified keys
